requirements.txt
streamlit
pillow
moviepy